Result component and formik on interest-rate.jsx
