+++ pages/styles/styles.css
@@ -851,12 +851,6 @@ video {
   outline-offset: 2px;
 }
 
-.transition-all {
-  transition-property: all;
-  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
-  transition-duration: 150ms;
-}
-
 .hover\:underline:hover {
   text-decoration-line: underline;
 }
